backend_fastapi_backup/tests/test_conversation_transfer.py
"""
Test suite for conversation transfer system
"""

import pytest
from uuid import UUID, uuid4
from datetime import datetime

from sqlalchemy.ext.asyncio import AsyncSession

from app.models.conversation import Conversation
from app.models.user import User
from app.models.department import Department
from app.services.conversation_service import ConversationService
from app.core.exceptions import NotFoundException


class TestTransferToAgent:
    """Tests for transfer_to_agent method"""

    @pytest.mark.asyncio
    async def test_transfer_to_agent_success(self, db: AsyncSession, org_id: UUID):
        """Test successful transfer to agent"""
        service = ConversationService(db)
        
        # Scenario: Transfer conversation to available agent
        # Should succeed and store history
        # TODO: Create test data fixtures
        # Expected: status='active', assigned_agent_id set, history recorded
        pass

    @pytest.mark.asyncio
    async def test_transfer_agent_without_capacity(self, db: AsyncSession, org_id: UUID):
        """Test transfer fails when agent is at capacity"""
        service = ConversationService(db)
        
        # Scenario: Agent already has max conversations
        # Should raise ValueError with capacity message
        # TODO: Create test data with agent at max capacity
        # Expected: Raises ValueError("Agent has reached maximum capacity...")
        pass

    @pytest.mark.asyncio
    async def test_transfer_agent_wrong_department(self, db: AsyncSession, org_id: UUID):
        """Test transfer fails when agent not in department"""
        service = ConversationService(db)
        
        # Scenario: Try to transfer to agent from different department
        # Should raise ValueError
        # TODO: Create test data with agents in different departments
        # Expected: Raises ValueError("Agent does not belong to conversation's department")
        pass

    @pytest.mark.asyncio
    async def test_transfer_agent_inactive(self, db: AsyncSession, org_id: UUID):
        """Test transfer fails when agent is inactive"""
        service = ConversationService(db)
        
        # Scenario: Target agent has is_active=false
        # Should raise ValueError
        # TODO: Create inactive user
        # Expected: Raises ValueError("Agent is not active")
        pass

    @pytest.mark.asyncio
    async def test_transfer_agent_unavailable_status(self, db: AsyncSession, org_id: UUID):
        """Test transfer fails when agent status not available"""
        service = ConversationService(db)
        
        # Scenario: Agent status is 'busy' or 'away'
        # Should raise ValueError
        # TODO: Create agent with agent_status='busy'
        # Expected: Raises ValueError("Agent is not available...")
        pass

    @pytest.mark.asyncio
    async def test_transfer_stores_history(self, db: AsyncSession, org_id: UUID):
        """Test transfer stores history in extra_data"""
        service = ConversationService(db)
        
        # Scenario: After successful transfer
        # Should have entry in extra_data["transfers"] with:
        # - from_agent_id
        # - to_agent_id
        # - note
        # - transferred_at (ISO format)
        # TODO: Verify history structure
        pass


class TestListAvailableAgents:
    """Tests for list_available_agents method"""

    @pytest.mark.asyncio
    async def test_list_available_agents_success(self, db: AsyncSession, org_id: UUID):
        """Test listing available agents"""
        service = ConversationService(db)
        
        # Scenario: List agents in department with capacity
        # Should return only agents with capacity_remaining > 0
        # TODO: Create multiple agents with different loads
        # Expected: Returns list with agents having capacity
        pass

    @pytest.mark.asyncio
    async def test_list_excludes_inactive_agents(self, db: AsyncSession, org_id: UUID):
        """Test that inactive agents are excluded"""
        service = ConversationService(db)
        
        # Scenario: Mix of active and inactive agents
        # Should exclude inactive (is_active=false)
        # TODO: Create mix of agent statuses
        # Expected: Only active agents returned
        pass

    @pytest.mark.asyncio
    async def test_list_excludes_full_capacity_agents(self, db: AsyncSession, org_id: UUID):
        """Test that agents at capacity are excluded"""
        service = ConversationService(db)
        
        # Scenario: Agents at max capacity
        # Should not appear in list
        # TODO: Create agent at max conversations
        # Expected: Agent not in returned list
        pass

    @pytest.mark.asyncio
    async def test_list_calculates_capacity_correctly(self, db: AsyncSession, org_id: UUID):
        """Test capacity_remaining calculation"""
        service = ConversationService(db)
        
        # Scenario: Agent with 5/10 conversations
        # Should show capacity_remaining=5
        # TODO: Create agent with specific load
        # Expected: capacity_remaining = max - active
        pass


class TestAssignWithDepartmentValidation:
    """Tests for assign_to_agent with department validation"""

    @pytest.mark.asyncio
    async def test_assign_agent_in_department_succeeds(self, db: AsyncSession, org_id: UUID):
        """Test assign succeeds when agent is in department"""
        service = ConversationService(db)
        
        # Scenario: Conversation has department, agent is in it
        # Should succeed
        # Expected: Conversation assigned, status='active'
        pass

    @pytest.mark.asyncio
    async def test_assign_agent_wrong_department_fails(self, db: AsyncSession, org_id: UUID):
        """Test assign fails when agent not in conversation's department"""
        service = ConversationService(db)
        
        # Scenario: Conversation assigned to Dept A, try assigning Dept B agent
        # Should raise ValueError
        # Expected: Raises ValueError("Agent does not belong to conversation's department")
        pass


# Integration test fixtures (pseudocode)
@pytest.fixture
async def conversation_in_queue(db: AsyncSession, org_id: UUID):
    """Create a queued conversation for testing"""
    # Create conversation with status='queued', assigned_department_id set
    # Return conversation_id
    pass


@pytest.fixture
async def agents_in_department(db: AsyncSession, org_id: UUID):
    """Create agents with various loads and statuses"""
    # Agent 1: available, 3/10 conversations
    # Agent 2: available, 10/10 conversations (at capacity)
    # Agent 3: busy, 2/10 conversations
    # Agent 4: inactive, 0/10 conversations
    # Return dict with agent_ids and metadata
    pass
