backend_fastapi_backup/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
motor>=3.6.0
slowapi>=0.1.9
httpx>=0.27.0
python-socketio>=5.11.0
python-dotenv>=1.0.0
redis>=5.0.0
pydantic>=2.6.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
cryptography>=42.0.0
passlib[argon2,bcrypt]>=1.7.4
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
tenacity>=8.2.0
celery>=5.3.0
PyJWT>=2.8.0
nest-asyncio>=1.6.0

# GraphQL
strawberry-graphql[fastapi]>=0.219.0

# Email
aiosmtplib>=3.0.0
Jinja2>=3.1.0

# Slack & HTTP
aiohttp>=3.9.0
strawberry-graphql[debug-server]>=0.219.0

# AI Providers
openai>=1.0.0
anthropic>=0.18.0
google-genai>=1.33.0

# SAML 2.0 / SSO
python3-saml>=1.15.0
authlib>=1.2.0

gevent==24.2.1
pyotp>=2.9.0
